third_party/WebKit/LayoutTests/dom/document_type/clone-node-expected.txt

Tests cloneNode for DocumentType.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.doctype is non-null.
PASS document.doctype.parentNode is document
PASS cd is non-null.
PASS cd.publicId is document.doctype.publicId
PASS cd.systemId is document.doctype.systemId
PASS cd.parentNode is null
PASS cd.ownerDocument is document
PASS cd.publicId is document.doctype.publicId
PASS cd.systemId is document.doctype.systemId
PASS cd.parentNode is null
PASS cd.ownerDocument is doc
PASS cd.parentNode is doc
PASS cd.name is 'a'
PASS cd.publicId is 'b'
PASS cd.systemId is 'c'
PASS cd.parentNode is null
PASS cd.ownerDocument is document
PASS successfullyParsed is true

TEST COMPLETE
